requirements.txt
Pillow==9.0.1
jupyter==1.0.0
jupytext==1.13.7
matplotlib==3.5.1
notebook==6.4.8
requests==2.27.1
cryptography
